Fix module shadowing issue by converting to absolute imports.
Fixes #103.
