+++ endpoints/openapi_generator.py
@@ -13,19 +13,20 @@
 # limitations under the License.
 
 """A library for converting service configs to OpenAPI (Swagger) specs."""
+from __future__ import absolute_import
 
 import json
 import logging
 import re
 import hashlib
 
-import api_exceptions
-import message_parser
+from . import api_exceptions
+from . import message_parser
 from protorpc import message_types
 from protorpc import messages
 from protorpc import remote
-import resource_container
-import util
+from . import resource_container
+from . import util
 
 
 _PATH_VARIABLE_PATTERN = r'{([a-zA-Z_][a-zA-Z_.\d]*)}'
