+++ endpoints/users_id_token.py
@@ -19,6 +19,8 @@ a user from an id_token.  The functionality provided by this library
 will be provided elsewhere in the future.
 """
 
+from __future__ import absolute_import
+
 import base64
 import json
 import logging
@@ -29,7 +31,7 @@ import urllib
 
 from collections import Container as _Container, Iterable as _Iterable
 
-import constants
+from . import constants
 
 from google.appengine.api import memcache
 from google.appengine.api import oauth
