+++ endpoints/util.py
@@ -15,6 +15,7 @@
 """Helper utilities for the endpoints package."""
 
 # pylint: disable=g-bad-name
+from __future__ import absolute_import
 
 import cStringIO
 import json
